Moved compilation tests to separate file
